# Faster count and apt package retriving method