--- a/src/lib.rs
+++ b/src/lib.rs
@@ -257,18 +257,12 @@ pub fn packages(manager: &str) -> io::Result<String> {
             let output = Command::new("apk").arg("info").output()?;
             Ok(format!("{}", packages::count(output)))
         }
-        "apt" => {
+        "apt" | "dpkg" => {
             let output = Command::new("dpkg").args(&["--get-selections"]).output()?;
             Ok(format!("{}", packages::count(output) - 1)) // -1 to deal with "Listing..."
         }
         "dnf" => {
             let output = Command::new("dnf").args(&["list", "installed"]).output()?;
-            Ok(format!("{}", packages::count(output)))
-        }
-        "dpkg" => {
-            let output = Command::new("dpkg-query")
-                .args(&["-f", "'${binary:Package}\n'", "-W"])
-                .output()?;
             Ok(format!("{}", packages::count(output)))
         }
         "eopkg" => {