--- a/src/packages.rs
+++ b/src/packages.rs
@@ -1,5 +1,4 @@
 pub fn count(output: std::process::Output) -> usize {
-    let raw_list = String::from_utf8_lossy(&output.stdout);
-    let list: Vec<&str> = raw_list.split('\n').collect();
-    list.iter().count() - 1 // -1 to deal with newline at end of output
-}
+    // -1 to deal with newline at end of output
+    String::from_utf8_lossy(&output.stdout).split('\n').count() - 1
+  }