--- a/src/packages.rs
+++ b/src/packages.rs
@@ -1,4 +1,4 @@
 pub fn count(output: std::process::Output) -> usize {
   // -1 to deal with newline at end of output
-  output.stdout.iter().filter(|&i| i == b'\n').count() - 1
+  output.stdout.iter().filter(|&&i| i == b'\n').count() - 1
 }