--- a/src/packages.rs
+++ b/src/packages.rs
@@ -1,4 +1,4 @@
 pub fn count(output: std::process::Output) -> usize {
   // -1 to deal with newline at end of output
-  String::from_utf8_lossy(&output.stdout).split('\n').count() - 1
+  output.stdout.iter().filter(|&&i| i == b'\n').count() - 1
 }