I noticed that retrieving total package count was slowing down the output and was quite jarring.
So I have proposed a new package retrieval method as well as new counting method that is quicker

| stock                        	| real    0.38s user    0.32s sys     0.06s cpu     100% 	|
|------------------------------	|--------------------------------------------------------	|
| dpkg w/ stock count          	| real    0.36s user    0.31s sys     0.06s cpu     99%  	|
| stock w/ new counting method 	| real    0.37s user    0.33s sys     0.04s cpu     99%  	|
| dpkg w/ new counting method  	| real    0.08s user    0.04s sys     0.04s cpu     97%  	|

These are the numbers I got during my test, I have not tried this on other systems other than a laptop and a pc