@Phate6660 2021-08-16
Just tried this out, the speed up is crazy!
I can merge this right away if you're done.
Thank you for the contribution. :)